Upgrade SharedStringTable
